Удалил лишние файлы Добавил поддержку WebAPI
