--- MyReceipts.UI/Ingridient.cs
@@ -1,28 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
-
-namespace Application;
-
-public class Ingridient
-{
-    private const int ZERO = 0;
-    
-    [Column("id")] 
-    public int Id { get; init; }
-
-    [Column("ingredient_name")] 
-    public string? Name { get; set; }
-
-
-    [Column("quantity")] 
-    public int Quantity { get; set; }
-
-
-    [Column("unit")] 
-    public string? Unit { get; set; }
-    
-    [Column("recipe_id")] 
-    public int RecipeId { get; set; }
-
-
-    public List<Recipe>? Recipes { get; init; }
-}
