+++ MyReceipts.UI/MyReceipts.UI.csproj
@@ -13,4 +13,9 @@
       <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.6" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\MyReceipts.App\MyReceipts.App.csproj" />
+      <ProjectReference Include="..\MyRecipts.WebApiHelperLib\MyRecipts.WebApiHelperLib.csproj" />
+    </ItemGroup>
+
 </Project>
