+++ MyReceipts.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyReceipts.App", "MyReceipt
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyReceipts.UI", "MyReceipts.UI\MyReceipts.UI.csproj", "{7342FB0E-7954-490E-AFDE-DABA1EEC1F45}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApiHelperLib1", "WebApiHelperLib1\WebApiHelperLib1.csproj", "{64B3CC2A-67FB-48E7-B039-7EC0BE7ADD28}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyRecipts.WebApiHelperLib", "MyRecipts.WebApiHelperLib\MyRecipts.WebApiHelperLib.csproj", "{511316A9-B382-40C6-A36C-844B8F0D2EB8}"
 EndProject
 Global
@@ -25,10 +23,6 @@ Global
 		{7342FB0E-7954-490E-AFDE-DABA1EEC1F45}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7342FB0E-7954-490E-AFDE-DABA1EEC1F45}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7342FB0E-7954-490E-AFDE-DABA1EEC1F45}.Release|Any CPU.Build.0 = Release|Any CPU
-		{64B3CC2A-67FB-48E7-B039-7EC0BE7ADD28}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{64B3CC2A-67FB-48E7-B039-7EC0BE7ADD28}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{64B3CC2A-67FB-48E7-B039-7EC0BE7ADD28}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{64B3CC2A-67FB-48E7-B039-7EC0BE7ADD28}.Release|Any CPU.Build.0 = Release|Any CPU
 		{511316A9-B382-40C6-A36C-844B8F0D2EB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{511316A9-B382-40C6-A36C-844B8F0D2EB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{511316A9-B382-40C6-A36C-844B8F0D2EB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
